tests/sonic_network_test.py
#!/usr/bin/env python3
"""
Sample pyATS test script for SONiC network topology testing

This script demonstrates basic connectivity testing for the SONiC network
topology using pyATS framework.
"""

import logging
import yaml
from pyats import aetest
from pyats.log.utils import banner
from pyats.topology import Testbed, loader

# Configure logging
logger = logging.getLogger(__name__)

# Global testbed object - will be initialized before tests run
global_testbed = None

class CommonSetup(aetest.CommonSetup):
    """Common setup tasks for all test cases"""
    
    @aetest.subsection
    def load_testbed(self, testbed):
        """Store the testbed object globally for all test methods"""
        global global_testbed
        
        logger.info(banner("Setting up global testbed object"))
        
        # Store testbed globally
        global_testbed = testbed
        
        logger.info(f"Testbed available with {len(testbed.devices)} devices")
        for device_name, device in testbed.devices.items():
            logger.info(f"  Device: {device_name} ({device.os})")
            
        self.passed(f"Testbed setup complete - {len(testbed.devices)} devices available")
    
    @aetest.subsection
    def check_topology(self):
        """Check if testbed is properly defined"""
        global global_testbed
        
        logger.info(banner("Checking testbed topology"))
        
        # Verify testbed is loaded
        assert global_testbed is not None, "Testbed not loaded"
        assert global_testbed.devices, "No devices found in testbed"
        
        # Log available devices
        for device_name, device in global_testbed.devices.items():
            logger.info(f"Found device: {device_name}")
            logger.info(f"  Type: {device.type}")
            logger.info(f"  OS: {device.os}")
            logger.info(f"  Connections: {list(device.connections.keys())}")
            
        self.passed(f"Testbed validation passed - found {len(global_testbed.devices)} devices")
    
    @aetest.subsection
    def test_device_connections(self):
        """Test connecting and disconnecting to all devices"""
        global global_testbed
        
        logger.info(banner("Testing device connections"))
        
        # Verify testbed is available
        assert global_testbed is not None, "Testbed not loaded"
        
        for device_name, device in global_testbed.devices.items():
            logger.info(f"Testing connection to {device_name}")
            
            # TODO: Implement device connection logic using the testbed object
            # Now you can use the actual device object for connections:
            # 1. device.connect() - Connect to the device using testbed credentials
            # 2. Verify connection is established
            # 3. Test basic command execution (e.g., device.execute('show version'))
            # 4. device.disconnect() - Disconnect from the device
            # 5. Verify connection is properly closed
            
            logger.info(f"  Device object available: {device}")
            logger.info(f"  Connections available: {list(device.connections.keys())}")
            logger.info(f"  TODO: device.connect() to establish connection")
            logger.info(f"  TODO: Test basic connectivity with device.execute()")
            logger.info(f"  TODO: device.disconnect() to close connection")
            
            # Placeholder assertion - replace with actual connection test
            self.passed(f"Connection test placeholder for {device_name}")

class SonicConnectivityTest(aetest.Testcase):
    """Test cases for SONiC network connectivity"""
    
    @aetest.setup
    def setup(self):
        """Setup for connectivity tests"""
        global global_testbed
        self.testbed = global_testbed
        assert self.testbed is not None, "Global testbed not available"
        
    @aetest.test
    def test_network_topology(self):
        """Test network topology connectivity"""
        logger.info(banner("Testing network topology"))
        
        # Define expected network links based on the topology
        expected_links = [
            ('sonic-1', 'sonic-2'),  # Path 1
            ('sonic-2', 'sonic-3'),
            ('sonic-3', 'sonic-4'),
            ('sonic-1', 'sonic-5'),  # Path 2
            ('sonic-5', 'sonic-6'),
            ('sonic-6', 'sonic-4'),
        ]
        
        for src, dest in expected_links:
            logger.info(f"Testing link: {src} <-> {dest}")
            # Add actual link connectivity test here
            # Example: ping test, LLDP verification, etc.
            self.passed(f"Link test passed: {src} <-> {dest}")
        
        # TODO: Implement data interface ping testing
        # Use the testbed.yaml file to discover actual IP addresses and test connectivity
        # Implementation hints:
        # 1. Parse testbed.topology.links to get interface mappings
        # 2. Extract IP addresses from device.custom.data_interfaces
        # 3. For each link, ping from source device to destination device's data interface IP
        # 4. Example: sonic-1 (192.1.2.11) should be able to ping sonic-2 (192.1.2.12)
        # 5. Use device.execute() or device.ping() methods to perform actual ping tests
        # 6. Verify bidirectional connectivity (both directions should work)
        # 7. Consider testing both direct neighbors and end-to-end paths
        
        logger.info("TODO: Implement ping tests between data interfaces")
        logger.info("  - Use testbed topology links to discover connections")
        logger.info("  - Extract IP addresses from device custom data_interfaces")
        logger.info("  - Test ping connectivity between connected interfaces")
        logger.info("  - Verify bidirectional connectivity")
        
        # Placeholder for now
        self.passed("Data interface ping testing - TODO implementation")

class CommonCleanup(aetest.CommonCleanup):
    """Common cleanup tasks"""
    
    @aetest.subsection
    def cleanup(self):
        """Cleanup after all tests"""
        logger.info(banner("Cleaning up test environment"))
        # Add any cleanup tasks here

if __name__ == '__main__':
    # Run the test
    import argparse
    import sys
    
    parser = argparse.ArgumentParser(description='SONiC Network Test Script')
    parser.add_argument('--testbed', '-t', 
                       help='Path to testbed YAML file',
                       default='testbed.yaml')
    
    args = parser.parse_args()
    
    # For standalone execution without testbed file
    if len(sys.argv) == 1:
        logger.info("Running basic test without testbed file")
        aetest.main()
    else:
        # Load testbed object from YAML file
        logger.info(f"Loading testbed from: {args.testbed}")
        testbed_obj = loader.load(args.testbed)
        logger.info(f"Testbed loaded with {len(testbed_obj.devices)} devices")
        
        # Run with testbed object
        aetest.main(testbed=testbed_obj)
