requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
arrow==1.3.0
async-lru==2.0.5
attrs==25.3.0
backports.ssl==0.0.9
backports.ssl_match_hostname==3.7.0.1
bcrypt==4.3.0
bidict==0.23.1
binaryornot==0.4.4
certifi==2025.8.3
cffi==2.0.0
chardet==4.0.0
charset-normalizer==3.4.3
ciscoisesdk==2.3.1
click==8.2.1
cookiecutter==2.6.0
coverage==7.10.6
cryptography==45.0.7
dict2xml==1.7.7
dill==0.4.0
distro==1.9.0
f5-icontrol-rest==1.3.13
fastjsonschema==2.21.2
frozenlist==1.7.0
genie==25.7
genie.libs.clean==25.7
genie.libs.conf==25.7
genie.libs.filetransferutils==25.7
genie.libs.health==25.7
genie.libs.ops==25.7
genie.libs.parser==25.7
genie.libs.robot==25.7
genie.libs.sdk==25.7
genie.telemetry==25.7
genie.trafficgen==25.7
gitdb==4.0.12
GitPython==3.1.45
grpcio==1.74.0
h11==0.16.0
idna==3.10
iniconfig==2.1.0
invoke==2.2.0
IxNetwork==9.30.2212.8
ixnetwork_restpy==1.8.0
Jinja2==3.1.6
jsonpickle==4.1.1
junit-xml==1.9
lxml==6.0.1
markdown-it-py==4.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
multidict==6.6.4
ncclient==0.7.0
netaddr==1.3.0
packaging==25.0
paramiko==4.0.0
pathspec==0.12.1
pluggy==1.6.0
ply==3.11
prettytable==3.16.0
propcache==0.3.2
protobuf==6.32.0
psutil==7.0.0
pyasn1==0.6.0
pyasynchat==1.0.4
pyasyncore==1.0.4
pyats==25.7
pyats.aereport==25.7
pyats.aetest==25.7
pyats.async==25.7
pyats.connections==25.7
pyats.contrib==25.7
pyats.datastructures==25.7
pyats.easypy==25.7
pyats.kleenex==25.7
pyats.log==25.7
pyats.reporter==25.7
pyats.results==25.7
pyats.robot==25.7
pyats.tcl==25.7
pyats.topology==25.7
pyats.utils==25.7
pycparser==2.22
pyftpdlib==2.0.1
Pygments==2.19.2
PyNaCl==1.5.0
pyOpenSSL==25.1.0
pysmi==1.6.2
pysnmp==6.1.4
pysnmpcrypto==0.0.4
pytest==8.4.2
pytest-cov==4.1.0
python-dateutil==2.9.0.post0
python-engineio==4.9.1
python-slugify==8.0.4
python-socketio==5.11.2
PyYAML==6.0.2
requests==2.32.5
requests-toolbelt==1.0.0
rest.connector==25.7
rich==14.1.0
robotframework==7.3.2
ruamel.yaml==0.18.15
ruamel.yaml.clib==0.2.12
setuptools==80.9.0
simple-websocket==1.1.0
six==1.17.0
smmap==5.0.2
text-unidecode==1.3
tftpy==0.8.0
tqdm==4.67.1
types-python-dateutil==2.9.0.20250822
unicon==25.7
unicon.plugins==25.7
urllib3==2.5.0
wcwidth==0.2.13
websocket-client==1.8.0
wheel==0.45.1
wsproto==1.2.0
xlrd==1.2.0
xlsxwriter==3.2.5
xlwt==1.3.0
xmltodict==0.12.0
yamllint==1.37.1
yang.connector==25.7
yarl==1.20.1
